+++ android/src/main/java/com/smileidentity/react/expo/SmileIDBiometricKYCView.kt
@@ -84,6 +84,7 @@ private fun BiometricKYCView(
             showAttribution = props.showAttribution,
             showInstructions = props.showInstructions,
             extraPartnerParams = props.extraParams,
+            consentInformation = props.consentInformation,
             idInfo = props.idInfo
         ) { result ->
             when(result) {
@@ -108,14 +109,7 @@ data class BiometricKYCProps(
     val skipApiSubmission: Boolean = false,
     val useStrictMode: Boolean = false,
     val extraParams: ImmutableMap<String, String> = persistentMapOf(),
-    val consentInformation: ConsentInformation = ConsentInformation(
-        consented = ConsentedInformation (
-            consentGrantedDate = getCurrentIsoTimestamp(),
-            personalDetails = false,
-            contactInformation = false,
-            documentInformation = false
-        )
-    ),
+    val consentInformation: ConsentInformation ? = null,
     val idInfo: IdInfo = IdInfo(
         country = ""
     )