+++ sample-expo/HomeScreen.tsx
@@ -1,35 +1,23 @@
-import React, { useEffect, useState } from 'react';
-import {
-  Alert,
-  FlatList,
-  SafeAreaView,
-  Text,
-  TouchableOpacity,
-  View,
-  StyleSheet,
-} from 'react-native';
+import React, {useEffect, useState} from 'react';
+import {Alert, FlatList, SafeAreaView, StyleSheet, Text, TouchableOpacity, View,} from 'react-native';
 
 import {
-  initialize,
-  setCallbackUrl,
-  setAllowOfflineMode,
-  submitJob,
-  getSubmittedJobs,
-  getUnsubmittedJobs,
-  cleanup,
-  SmileIDDocumentVerificationView,
-  SmileIDSmartSelfieEnrollmentView,
-  SmileIDDocumentVerificationEnhancedView,
-  SmileConfig,
+  AutoCapture,
+  BiometricKYCParams,
+  ConsentInformationParams,
   DocumentVerificationParams,
   EnhancedDocumentVerificationParams,
+  IdInfoParams,
+  initialize,
   SmartSelfieParams,
-  SmileIDSmartSelfieAuthenticationEnhancedView,
-  SmileIDSmartSelfieEnrollmentEnhancedView,
-  SmileIDSmartSelfieAuthenticationView,
+  SmileConfig,
   SmileIDBiometricKYCView,
-  BiometricKYCParams,
-  IdInfoParams, ConsentInformationParams,
+  SmileIDDocumentVerificationEnhancedView,
+  SmileIDDocumentVerificationView,
+  SmileIDSmartSelfieAuthenticationEnhancedView,
+  SmileIDSmartSelfieAuthenticationView,
+  SmileIDSmartSelfieEnrollmentEnhancedView,
+  SmileIDSmartSelfieEnrollmentView,
 } from 'react-native-expo';
 
 import DocumentVerificationEnhancedSvgIcon from "./icons/DocumentVerificationEnhancedSvgIcon";
@@ -92,9 +80,10 @@ const documentVerificationParams: DocumentVerificationParams = {
   countryCode: 'NG',
   allowNewEnroll: false,
   documentType: 'PASSPORT',
-  // idAspectRatio: 1.414, // Optional aspect ratio for document capture
+  // idAspectRatio: 1.0, // Optional aspect ratio for document capture
   // bypassSelfieCaptureWithFile: '', // Optional file path to bypass selfie capture
-  enableAutoCapture: true,
+  autoCaptureTimeout: 10, // this is in seconds,
+  autoCapture: AutoCapture.AutoCapture,
   captureBothSides: false,
   allowAgentMode: true,
   showInstructions: true,
@@ -136,9 +125,10 @@ const enhancedDocumentVerificationParams: EnhancedDocumentVerificationParams = {
   countryCode: 'NG',
   allowNewEnroll: false,
   documentType: 'PASSPORT',
-  // idAspectRatio: 1.414, // Optional aspect ratio for document capture
+  // idAspectRatio: 1.0, // Optional aspect ratio for document capture
   // bypassSelfieCaptureWithFile: '', // Optional file path to bypass selfie capture
-  enableAutoCapture: false,
+  autoCaptureTimeout: 10, // this is in seconds,
+  autoCapture: AutoCapture.AutoCapture,
   captureBothSides: false,
   allowGalleryUpload: true,
   showInstructions: true,
@@ -149,7 +139,7 @@ const enhancedDocumentVerificationParams: EnhancedDocumentVerificationParams = {
     'custom_param_1': 'value1',
     'custom_param_2': 'value2'
   },
-  consentInformation: consentInformationParams
+  consentInformation: consentInformationParams // Optional consent information
 };
 
 
@@ -178,7 +168,7 @@ const biometricKYCParams: BiometricKYCParams = {
     'custom_param_1': 'value1',
     'custom_param_2': 'value2'
     },
-  consentInformation: consentInformationParams,
+  consentInformation: consentInformationParams, // Optional consent information
   idInfo: idInfoParams
 };