+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## 11.1.0
+
+* Changed `enableAutoCapture` to AutoCapture enum to allow partners change document capture options
+* Added `autoCaptureTimeout` to allow partners to set a timeout for auto capture
+* Updated Smile ID android SDK version to 11.1.0
+* Bumped `compileSdkVersion` to 36 and fixed build issues
+* Removed default `ConsentInformation` values as the new SDK allows it to be nullable
+
 ## 11.0.0 - July 31, 2025
 
 * Initial Release